DATAMONGO-843 -  Unable to use @EnableMongoAuditing annotation in Java config.
Another round for DATAMONGO-843.

Changed BeanNames.MAPPING_CONTEXT constant to "mongoMappingContext" to explicitly reference the MongoMappingContext of AbstractMongoConfiguration.mongoMappingContext(). Adjusted tests accordingly.
Previously we just referred to "mappingContext" which did not match our default bean name for the MongoMappingContext.

We now use an ObjectFactory as the constructor argument in order to avoid a dependency cycle. This cycle arises from the initialisation of the MappingContext (afterPropertiesSet()) triggers the bean creation of the AuditingEventListener due to the eager initialisation of ApplicationEventListeners by the spring framework. This implicitly requests the bean creation of the MappingContextIsNewStrategy which requires the initial MappingContext that is still in construction which forms the cycle.